Ignore any prefix that suggest to be a file.
Ignore any prefix that suggest to be a file followed by a colon, as
reported by the output of grep.

For instance, consider what is returned by the following command:

  $ fgrep -r bash .
  bin/hd-env-bashrc-install:# On RHEL systems..
  bin/hd-env-bashrc-install:if fgrep -q '.bashrc' \"\\\${HOME}/.bash_profile\"; then
  bin/hd-env-bashrc-install:  BASH_FILE=\"\\\${HOME}/.bash_profile\"

Now notice that when passed through '${SCRIPT_NAME} -f', the line with
the comments does not show because the it is commented, even though the
actual line in <STDIN> does not start with a comment, but with a
filename, i.e. 'bin/hd-env-bashrc-install:'.

  $ fgrep -r bash . | ${SCRIPT_NAME} -f
  bin/hd-env-bashrc-install:if fgrep -q '.bashrc' \"\\\${HOME}/.bash_profile\"; then
  bin/hd-env-bashrc-install:  BASH_FILE=\"\\\${HOME}/.bash_profile\"

This option is ignored when input is XML.