Add mapping debugging code, commented out.
Add mapping debugging code but commented out.  Introduce new function
print_mappings().

print_mappings() is a debugging function that prints the hexadecimal
address of the function, not the name.  Apparently, Lua does not save
the name of functions, so the utility of this function is limited.  This
is why it is commented out, but maybe one day it could be useful.