Moved vim/nvim code to "env.gen.noarch.bash'.
Moved vim/nvim code from 'sys.gen.noarch.dash' to "env.gen.noarch.bash'.
The code is to sophisticated to be justified to find itself in a sys.*
file which is meant to be minimalistic for bear systems.

This also fix an issue when nvim's path was net yet set, being added in
a subsequent environment file, as 'sys.gen.noarch.dash' is being loaded
very early in the boot sequence.  Since nvim's path was not set, all
nvim aliases were not being loaded.